Add support for a new passwordeval configuration directive
This directive will allow to produce the cleartext password from an
external command during runtime. Thus the cleartext password needs not
to be stored in cleartext in the config file. This approach does not
require special interfaces to other programs and the user can choose
whichever program he wants that allows to acquire the cleartext password
from stdin.

A simple usage is to call gpg to decrypt the password from an encrypted
file. If a gpg-agent is running then the password for a gpg keyring will
be interactively queried to unlock it.

For easy testing we can also use something like:

  passwordeval "cat /my/file"

to obtain the password in cleartext from some file.

If a cleartext password is still given in the configuration then it
takes precedence over the passwordeval.