socket.c: fix catastrophic brace mistake
In SSLOpen(), it looks like a merge resulted in some code getting mis-indented
such that the initialization of `global_mydata_index` only occurred in the
unlikely and undesirable event that the runtime OpenSSL library version
disagreed with the compile-time version. This basically makes the program
segfault reliably if SSL is used.

Fix this, and reindent this section to remove the counterintuitive bits.