fetchmailconf.py: hostname qualification fixup
If socket.gethostname() returns a qualified name, don't look further,
to match fetchmail's behaviour - in case of an FQDN /etc/hostname
with broken /etc/hosts, this helps us survive.

Else, call socket.getfqdn() and not gethostbyaddr() in an attempt to qualify
the hostname.

Failing that, print an error message that tells the user to fix
/etc/hosts, which has this canonical format:
10.9.8.7 host.example.org host
on the assumption that that line and DNS might both be broken.

Gitlab, fixes #12 reported by Sergey Alirzaev - the prior attempt fixed
the wrong place (which also needed fixing).

Bump version to 1.63.