Add an option to force idle.
I noticed some servers do support IDLE mode but don't advertise it
correctly. It's the case for free.fr servers (IDLE is advertised in the
initial HELO, but not on explicit capa probe). Also the case for iCloud
which does support it (iCloud is using dovecot AFAIU). Use this flag on
top of --idle to override what the server says.